Module ran::generators

The low level type specific generators

Constants

MANTISSA_MAX

Constant for converting u64 numbers to f64s in [0,1). It is the maximum value of mantissa plus one.

SEED

SEED is used by ranf64 and/or splitmix algorithms. It is initialised by default to systime seconds.

Functions

get_seed

For saving the current value of seed, reproducing the same sequence later

get_xoshi

Load the xoshiro seeds into an array so as not to have to pass them round everywhere as an &mut [u64;4] argument

put_xoshi

Put the xoshiro seeds back from a passed array

ran_frange

Transform an f64 number in [0,1) to the range [min:f64,max:f64)

ran_i64

i64 random number by simply casting from xoshiu64

ran_irange

Generate i64 random number in the interval [min:i64,max:i64], where the interval may span zero. Must be min<max always!

ran_ubits

Get random numbers of various smaller unsigned integer types by specifying the number of bits required,
e.g. ran_ubits(16) as u16, etc.

ran_urange

Generate u64 random number in the interval [min,max]. You can recast the result into some smaller type, when you know that it will fit in.

ranf64

Generate f64 random number in the standardised range [0,1). It can be linearly transformed to any [min,max] range.
Very fast, using just three shift and XOR instructions. Based on: George Marsaglia, Xorshift RNGs, Journal of Statistical Software 08(i14), Jan 2003. Disclaimer: for cryptography, use real random source.

ranvf64

Generates vector of size d, of f64 random numbers in [0,1).

ranvf64_xoshi

Generates vector of size d, of f64 random numbers in [0,1). Bit slower but otherwise superior to ranvf64.

ranvi64

Generates vector of size d, of i64 random numbers.

ranvi64_in

Generates vector of size d, of i64 random numbers in the interval [min,max]. May include zero.

ranvu8

Generates vector of size d, of u8 random numbers in [0,255]. You can similarly recast u64 yourself to any other type.

ranvu16

Generates vector of size d, of u16 random numbers in [0,65535]. You can similarly recast u64 yourself to any other type.

ranvu64

Generates vector of size d, filled with full range u64 random numbers.

ranvvf64

Generates n vectors of size d each, of f64 random numbers in [0,1).

ranvvf64_xoshi

Generates n vectors of size d each, of f64 random numbers in [0,1).

ranvvi64

Generates n vectors of size d each, of i64 random numbers in the interval [min,max].

ranvvi64_in

Generates n vectors of size d each, of i64 random numbers in the interval [min,max].

ranvvu8

Generates n vectors of size d each, of u8 random numbers in the interval [0,255].

ranvvu16

Generates n vectors of size d each, of u16 random numbers in the interval [0,65535].

ranvvu64

Generates n vectors of size d each, of full range u64 random numbers.

set_seeds

Manual initialisation of SEED (and derived xoshi seeds). When seed == 0, resets SEED to systime in nanoseconds, producing an essentially unpredictable random sequence. Otherwise sets the SEED to the argument value. The latter will repeat the same unique sequence for each value.

splitmix

Simple SPLITMIX64 fast generator recommended for generating the initial sequence of seeds. Assumes that SEED has been set and uses it. It can generate any sequence of random u64s. Passes the BigCrush test. Adapted from Sebastiano Vigna, 2015.

xoshi_step

Core part of the xoshi (xor shift) algorithms

xoshif64

Possibly the best f64 generator. Translated and modified from: xoshiro256+1.0 algorithm by David Blackman and Sebastiano Vigna (vigna@acm.org), 2018. Also added conversion to f64 output in the range [0,1)

xoshiu64

Possibly the best full 64 bits generator. Adapted from xoshiro256** 1.0 algorithm by David Blackman and Sebastiano Vigna (vigna@acm.org), 2018.