Crate ran

Mean, lean, fast generation of random numbers of various types. Also filling vectors and vectors of vectors with random numbers

Re-exports

pub use crate::generators::set_seeds;

pub use crate::generators::get_seed;

pub use crate::error::RanError;

pub use crate::error::rerror;

pub use crate::error::Re;

Modules

error

Custom error RError

generators

The low level type specific generators

impls

Implementations of associated functions for the enum types

secondary

Some helper functions

Enums

Rnum

Wrapper for enum polymorphism - supported end types

Rv

Wrapper for enum polymorphism - vector

Rvv

Wrapper for enum polymorphism - vector of vectors